fixes unit test cases
